+++ tests/test_pytypes.py
@@ -120,23 +120,17 @@ def test_str(doc):
     assert s1 == s2
 
     malformed_utf8 = b"\x80"
+    if hasattr(m, "PYBIND11_STR_LEGACY_PERMISSIVE"):
+        assert m.str_from_object(malformed_utf8) is malformed_utf8
+    elif env.PY2:
+        with pytest.raises(UnicodeDecodeError):
+            m.str_from_object(malformed_utf8)
+    else:
+        assert m.str_from_object(malformed_utf8) == "b'\\x80'"
     if env.PY2:
-        if hasattr(m, "has_str_non_permissive"):
-            with pytest.raises(UnicodeDecodeError):
-                m.str_from_object(malformed_utf8)
-        else:
-            m.str_from_object(
-                malformed_utf8
-            ) is malformed_utf8  # To be fixed; see #2380
         with pytest.raises(UnicodeDecodeError):
             m.str_from_handle(malformed_utf8)
     else:
-        if hasattr(m, "has_str_non_permissive"):
-            assert m.str_from_object(malformed_utf8) == "b'\\x80'"
-        else:
-            assert (
-                m.str_from_object(malformed_utf8) is malformed_utf8
-            )  # To be fixed; see #2380
         assert m.str_from_handle(malformed_utf8) == "b'\\x80'"
 
 
@@ -312,26 +306,26 @@ def test_pybind11_str_raw_str():
     valid_orig = u"Ǳ"
     valid_utf8 = valid_orig.encode("utf-8")
     valid_cvt = cvt(valid_utf8)
-    if hasattr(m, "has_str_non_permissive"):
+    if hasattr(m, "PYBIND11_STR_LEGACY_PERMISSIVE"):
+        assert valid_cvt is valid_utf8
+    else:
         assert type(valid_cvt) is unicode if env.PY2 else str  # noqa: F821
         if env.PY2:
             assert valid_cvt == valid_orig
         else:
-            assert valid_cvt == u"b'\\xc7\\xb1'"
-    else:
-        assert valid_cvt is valid_utf8
+            assert valid_cvt == "b'\\xc7\\xb1'"
 
     malformed_utf8 = b"\x80"
-    if hasattr(m, "has_str_non_permissive"):
+    if hasattr(m, "PYBIND11_STR_LEGACY_PERMISSIVE"):
+        assert cvt(malformed_utf8) is malformed_utf8
+    else:
         if env.PY2:
             with pytest.raises(UnicodeDecodeError):
                 cvt(malformed_utf8)
         else:
             malformed_cvt = cvt(malformed_utf8)
-            assert type(malformed_cvt) is unicode if env.PY2 else str  # noqa: F821
-            assert malformed_cvt == u"b'\\x80'"
-    else:
-        assert cvt(malformed_utf8) is malformed_utf8
+            assert type(malformed_cvt) is str
+            assert malformed_cvt == "b'\\x80'"
 
 
 def test_implicit_casting():
@@ -517,10 +511,10 @@ def test_isinstance_string_types():
     assert not m.isinstance_pybind11_bytes(u"")
 
     assert m.isinstance_pybind11_str(u"")
-    if hasattr(m, "has_str_non_permissive"):
-        assert not m.isinstance_pybind11_str(b"")
-    else:
+    if hasattr(m, "PYBIND11_STR_LEGACY_PERMISSIVE"):
         assert m.isinstance_pybind11_str(b"")
+    else:
+        assert not m.isinstance_pybind11_str(b"")
 
 
 def test_pass_bytes_or_unicode_to_string_types():
@@ -528,23 +522,22 @@ def test_pass_bytes_or_unicode_to_string_types():
     with pytest.raises(TypeError):
         m.pass_to_pybind11_bytes(u"Str")
 
-    if hasattr(m, "has_str_caster_no_implicit_decode"):
+    if hasattr(m, "PYBIND11_STR_LEGACY_PERMISSIVE") or env.PY2:
+        assert m.pass_to_pybind11_str(b"Bytes") == 5
+    else:
         with pytest.raises(TypeError):
             m.pass_to_pybind11_str(b"Bytes")
-    else:
-        assert m.pass_to_pybind11_str(b"Bytes") == 5
     assert m.pass_to_pybind11_str(u"Str") == 3
 
     assert m.pass_to_std_string(b"Bytes") == 5
     assert m.pass_to_std_string(u"Str") == 3
 
     malformed_utf8 = b"\x80"
-    if hasattr(m, "has_str_non_permissive"):
-        if hasattr(m, "has_str_caster_no_implicit_decode"):
-            with pytest.raises(TypeError):
-                m.pass_to_pybind11_str(malformed_utf8)
-        else:
-            with pytest.raises(UnicodeDecodeError):
-                m.pass_to_pybind11_str(malformed_utf8)
-    else:
+    if hasattr(m, "PYBIND11_STR_LEGACY_PERMISSIVE"):
         assert m.pass_to_pybind11_str(malformed_utf8) == 1
+    elif env.PY2:
+        with pytest.raises(UnicodeDecodeError):
+            m.pass_to_pybind11_str(malformed_utf8)
+    else:
+        with pytest.raises(TypeError):
+            m.pass_to_pybind11_str(malformed_utf8)